feature: credito 1 funcionando

+++ src/main/java/com/example/finanzasgrupo5backend/Credito1/Controller/Credito1Controller.java
@@ -51,7 +51,7 @@ public class Credito1Controller {
     @PatchMapping("/clienteId/{id}")
     public ResponseEntity<Credito1Response> updateCredito1(@PathVariable(name = "id") Long creditoId,
                                                            @RequestParam LocalDate fechaInicial, @RequestParam LocalDate fechaFinal, @RequestParam String TEoN,@RequestParam String TEP, @RequestParam String TNP,
-                                                           @RequestParam Long tasa, @RequestParam String perio_capitalizacion) {
+                                                           @RequestParam Double tasa, @RequestParam String perio_capitalizacion) {
 
         long clienteId=1;
         var res = credito1Service.updateCredito1(creditoId, fechaInicial, fechaFinal, TEoN,TEP, TNP,
