+++ src/main/java/com/example/finanzasgrupo5backend/Credito1/Model/Credito1Response.java
@@ -19,7 +19,7 @@ public class Credito1Response {
     private String TEoN;
     private String TEP;
     private String TNP;
-    private Long tasa;
+    private Double tasa;
     private String perio_capitalizacion;
 
     //private Client client;
