+++ src/main/java/com/example/finanzasgrupo5backend/Credito1/Model/MoraCredito1Response.java
@@ -13,9 +13,9 @@ public class MoraCredito1Response {
 
     private Long id;
     private String TEPm;
-    private Long tasa;
+    private Double tasa;
     private Long dias_atraso;
-    private Long total_moras;
+    private Double total_moras;
 
     //private Credito1 credito1;
 }
