+++ src/main/java/com/example/finanzasgrupo5backend/Credito1/Service/ICredito1Service.java
@@ -21,7 +21,7 @@ public interface ICredito1Service {
     Credito1Response createCredito1(Credito1Request credito1Request, Long clienteId);
 
     public abstract Credito1Response updateCredito1(Long id, LocalDate fechaInicial, LocalDate fechaFinal, String TEoN,String TEP, String TNP,
-                                                    Long tasa, String perio_capitalizacion, Long clienteId);
+                                                    Double tasa, String perio_capitalizacion, Long clienteId);
 
     public abstract Credito1Response deleteCredito1(Long id);
